requirements.txt
easydict==1.10
einops==0.5.0
numpy==1.22.4
pandas==1.4.0
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
torch==1.10.0
torchvision==0.11.1
vit-pytorch==0.35.8
pytorch-metric-learning==1.6.3